+++ apistar_peewee_orm/__init__.py
@@ -1,5 +1,5 @@
 from apistar_peewee_orm.components import PeeweeDatabaseComponent
-from apistar_peewee_orm.event_hooks import PeeweeTransactionHook
 from apistar_peewee_orm.database import Model
+from apistar_peewee_orm.event_hooks import PeeweeTransactionHook
 
 __all__ = ["PeeweeDatabaseComponent", "PeeweeTransactionHook", "Model"]